confirmed working multi-letter elements, and cleaned up the element inheritance so that addition of a new element only requires adding it to the file.  The function names are added to a dictionary where the function call is the key, i.e. R(a,b) is represented as 'R'.  Downstream code required no modification, as these were added to the globals() dictionary upon import. > The tests were updated so that the inputs and expected outputs are held in dictionaries so that order is no longer important. > The code to extract the elements from the circuit was altered significantly, and a test of this function has been added.  It has been demonstrated to support arbitrarily high integers and still properly parses them. The new element, RR, is a resistor whose value is multiplied by 100.  I've verified that, when fit, the parameter value is 100-fold smaller than the expeccted R value, and that the fits are identical.
